+++ protocols/floodsub/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 - Update dependencies.
 
+- Change `FloodsubDecodeError::ReadError` from a `upgrade::ReadOneError` to
+  `std::io::Error`. See [PR 2111].
+
+[PR 2111]: https://github.com/libp2p/rust-libp2p/pull/2111
+
 # 0.29.0 [2021-04-13]
 
 - Update `libp2p-swarm`.